requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
django-filter==25.1
djoser==2.3.1
sqlparse==0.5.3
tzdata==2025.2
pillow==11.2.1
